unit: fix line length check to deal with values
If an option value is split across multiple lines, the current
line-length check does not catch it as it only considers lines from the
start of them. Hence, we add an additional check while parsing an
option value.

This is not necessary in section headers or option names because they are
alway checked in the core next() loop; it affects values because the
lexer recurses to parse them.

I (and everyone else) missed this while implementing this check
originally because the lexing code is way too complex.